requirements.txt
# Core orchestration
langchain>=0.3.0
langgraph>=0.3.0
langchain-openai>=0.3.0

# Config and data
pydantic>=2.7.0
python-dotenv>=1.0.1
email-validator>=2.2.0

# Scheduling
apscheduler>=3.10.0
pytz>=2023.3

# Backend API
fastapi>=0.112.0
uvicorn>=0.30.0
motor>=3.5.0
PyJWT>=2.9.0
python-multipart>=0.0.6

# PDF and document processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.1.0

# Database
pymongo>=4.6.0

# AI/ML
google-generativeai>=0.3.0

# Email service
sendgrid>=6.11.0